refactor: reservation 영속 계층 패키지명 변경 및 ReservationStatus factory method 예외 변경
